Refactor code to pull user specific logic out of PlayerPanel
Used the NullPlayer class to differentiate between regular players and
non-logged players.  Used this to place specific items on the
PlayerPanel such as the title of the panel, and the buttons
"createAccount", "login", and "logout".  This way the logic is
controlled by the PlayerOnePanel and PlayerTwoPanel classes and not by
the PlayerPanel class.